+++ lib/dx/defd/result.ex
@@ -103,6 +103,13 @@ defmodule Dx.Defd.Result do
     |> transform(&Enum.reverse/1)
   end
 
+  def reject(enum, fun, eval) do
+    Enum.reduce_while(enum, ok([]), fn elem, acc ->
+      combine(:reject, acc, fun.(elem, eval), elem)
+    end)
+    |> transform(&Enum.reverse/1)
+  end
+
   @doc """
   Applies the `mapper` function to each element and reduces the mapped results
   using `acc` and `fun` as long as the mapped results are `{:ok, _}` tuples.
@@ -509,6 +516,11 @@ defmodule Dx.Defd.Result do
   defp combine(:filter, acc, {:ok, false}, _), do: {:cont, acc}
   defp combine(:filter, acc, {:ok, nil}, _), do: {:cont, acc}
   defp combine(:filter, acc, {:ok, _}, elem), do: combine(:all, acc, {:ok, elem}, & &1)
+
+  # :reject
+  defp combine(:reject, acc, {:ok, false}, elem), do: combine(:all, acc, {:ok, elem}, & &1)
+  defp combine(:reject, acc, {:ok, nil}, elem), do: combine(:all, acc, {:ok, elem}, & &1)
+  defp combine(:reject, acc, {:ok, _}, _), do: {:cont, acc}
 
   # :any?
   defp combine(:any?, _acc, {:ok, true}, _), do: {:halt, {:ok, true}}