+++ lib/dx/ecto/scope.ex
@@ -22,6 +22,8 @@ defmodule Dx.Ecto.Scope do
     post_load: {:loaded}
   }
 
+  defguard in_subquery?(state) when length(state.queries) > 1
+
   def to_query(_queryable, %{scope: scope}) do
     state = %{@state | post_load: scope.post_load}
 
@@ -123,10 +125,24 @@ defmodule Dx.Ecto.Scope do
     {{:count, base}, nil, refs}
   end
 
+  defp resolve({:any?, base, fallback}, refs) do
+    {base, _ref, refs} = resolve(base, refs)
+    {{:any?, base, fallback}, nil, refs}
+  end
+
   defp resolve({:filter, base, condition}, refs) do
     {base, ref, refs} = resolve(base, refs)
     {condition, _ref, refs} = resolve_condition(condition, ref, refs)
     {{:filter, base, condition}, ref, refs}
+  end
+
+  defp resolve({:compare, left, right, fallback}, refs) do
+    with {left, _ref, refs} <- resolve(left, refs),
+         {right, _ref, refs} <- resolve(right, refs) do
+      {{:compare, left, right}, nil, refs}
+    else
+      :error -> {{:error, fallback}, nil, refs}
+    end
   end
 
   defp resolve(value, refs) do
@@ -158,6 +174,16 @@ defmodule Dx.Ecto.Scope do
     {{:all_of, conditions}, ref, refs}
   end
 
+  defp resolve_condition({:any_of, conditions, fallback}, ref, refs) do
+    {conditions, refs} =
+      Enum.map_reduce(conditions, refs, fn condition, refs ->
+        {condition, _ref, refs} = resolve_condition(condition, ref, refs)
+        {condition, refs}
+      end)
+
+    {{:any_of, conditions, fallback}, ref, refs}
+  end
+
   defp resolve_condition({:error, fallback}, ref, refs) do
     {{:error, fallback}, ref, refs}
   end
@@ -173,6 +199,14 @@ defmodule Dx.Ecto.Scope do
 
   defp resolve_condition({:&&, condition, then, fallback}, ref, refs) do
     resolve_condition({:and, condition, then, fallback}, ref, refs)
+  end
+
+  defp resolve_condition({:or, left, right, fallback}, ref, refs) do
+    resolve_condition({:any_of, [left, right], fallback}, ref, refs)
+  end
+
+  defp resolve_condition({:||, condition, then, fallback}, ref, refs) do
+    resolve_condition({:or, condition, then, fallback}, ref, refs)
   end
 
   defp resolve_condition({:eq, {:error, _fallback}, _right, fallback}, ref, refs) do
@@ -191,10 +225,12 @@ defmodule Dx.Ecto.Scope do
     {{:error, fallback}, ref, refs}
   end
 
-  defp resolve_condition({:eq, left, right, fallback}, ref, refs) do
+  @ops [:lt, :eq, :gt]
+  defp resolve_condition({op, left, right, fallback}, ref, refs)
+       when op in @ops do
     with {left, _ref, refs} <- resolve(left, refs),
          {right, _ref, refs} <- resolve(right, refs) do
-      {{:eq, left, right}, ref, refs}
+      {{op, left, right}, ref, refs}
     else
       :error -> {{:error, fallback}, ref, refs}
     end
@@ -207,14 +243,33 @@ defmodule Dx.Ecto.Scope do
   # NORMALIZE
   # ---------
 
+  defp normalize({:filter, {:filter, base, condition}, other_condition}) do
+    normalize({:filter, base, {:all_of, [condition, other_condition]}})
+  end
+
   defp normalize({:filter, base, condition}) do
     {:filter, normalize(base), normalize_condition(condition)}
   end
 
   defp normalize(other), do: other
 
+  defp flatten_condition(op, conditions) do
+    Enum.flat_map(conditions, fn
+      {^op, conditions} -> flatten_condition(op, conditions)
+      condition -> [normalize_condition(condition)]
+    end)
+  end
+
   defp normalize_condition({:all_of, conditions}) do
-    {:all_of, Enum.map(conditions, &normalize_condition/1)}
+    {:all_of, flatten_condition(:all_of, conditions)}
+  end
+
+  defp normalize_condition({:any_of, conditions, fallback}) do
+    {:any_of, flatten_condition(:any_of, conditions), fallback}
+  end
+
+  defp normalize_condition({:any?, base, fallback}) do
+    {:any?, normalize(base), fallback}
   end
 
   defp normalize_condition({:not, condition}) do
@@ -226,15 +281,48 @@ defmodule Dx.Ecto.Scope do
     |> normalize_condition()
   end
 
-  # assoc == %struct{}
+  # Ecto.Enum
+  defp normalize_condition(
+         {:eq, left,
+          {:field, {:parameterized, Ecto.Enum, %{type: type, mappings: mappings}}, base, field}}
+       ) do
+    {:eq, {:field, {:parameterized, Ecto.Enum, %{type: type, mappings: mappings}}, base, field},
+     left}
+    |> normalize_condition()
+  end
+
   defp normalize_condition(
          {:eq,
+          {:field, {:parameterized, Ecto.Enum, %{type: type, mappings: mappings}}, base, field},
+          right}
+       ) do
+    value =
+      Keyword.get(mappings, right) ||
+        raise ArgumentError,
+              "#{inspect(right)} is not a valid value for field #{field}. Valid values are: " <>
+                Enum.map_join(mappings, ", ", &inspect(elem(&1, 0)))
+
+    {:eq, {:field, type, base, field}, value}
+    |> normalize_condition()
+  end
+
+  # assoc == %struct{}
+  defp normalize_condition(
+         {op,
           {:as, _, type,
            {:assoc, :one, foreign_key_type, foreign_key_field, _related_key_type,
             related_key_field, base, _field}}, %type{} = struct}
-       ) do
-    {:eq, {:field, foreign_key_type, normalize(base), foreign_key_field},
+       )
+       when op in @ops do
+    {op, {:field, foreign_key_type, normalize(base), foreign_key_field},
      Map.fetch!(struct, related_key_field)}
+    |> normalize_condition()
+  end
+
+  # compare/2 functions
+  defp normalize_condition({:eq, {:compare, left, right}, op})
+       when op in [:lt, :gt, :eq] do
+    {op, left, right}
     |> normalize_condition()
   end
 
@@ -301,14 +389,18 @@ defmodule Dx.Ecto.Scope do
   end
 
   defp build({:as, new_ref, type, {:assoc, :one, _, _, _, _, base, field}}, state) do
-    {{:ref, ref}, state} = build(base, state)
+    case build(base, state) do
+      {{:ref, ref}, state} ->
+        state =
+          map_query(state, &join(&1, :inner, [{^ref, x}], y in assoc(x, ^field), as: ^new_ref))
+          |> Map.put(:alias_types, Map.put(state.alias_types, new_ref, type))
 
-    state =
-      map_query(state, &join(&1, :inner, [{^ref, x}], y in assoc(x, ^field), as: ^new_ref))
-      |> Map.put(:alias_types, Map.put(state.alias_types, new_ref, type))
+        {:ref, new_ref}
+        |> with_state(state)
 
-    {:ref, new_ref}
-    |> with_state(state)
+      :error ->
+        :error
+    end
   end
 
   defp build(
@@ -331,16 +423,25 @@ defmodule Dx.Ecto.Scope do
     |> with_state(state)
   end
 
+  defp build({:field, _type, {:ref, ref}, field}, %{queries: [%{ref: ref} | _]} = state) do
+    dynamic(field(as(^ref), ^field))
+    |> with_state(state)
+  end
+
   defp build({:field, _type, {:ref, ref}, field}, state) do
-    dynamic([{^ref, x}], field(x, ^field))
+    dynamic(field(parent_as(^ref), ^field))
     |> with_state(state)
   end
 
   defp build({:field, _type, base, field}, state) do
-    {{:ref, ref}, state} = build(base, state)
+    case build(base, state) do
+      {{:ref, ref}, state} ->
+        dynamic([{^ref, x}], field(x, ^field))
+        |> with_state(state)
 
-    dynamic([{^ref, x}], field(x, ^field))
-    |> with_state(state)
+      :error ->
+        :error
+    end
   end
 
   defp build({:count, base}, state) do
@@ -350,12 +451,21 @@ defmodule Dx.Ecto.Scope do
   end
 
   defp build({:filter, base, condition}, state) do
-    {ref, state} = build(base, state)
-    {condition, state} = build_condition(condition, ref, state)
+    case build(base, state) do
+      {ref, state} ->
+        case build_condition(condition, state) do
+          {condition, fallback, state} ->
+            state = state |> build_query_filter(condition) |> build_fallback_filter(fallback)
 
-    state = map_query(state, &where(&1, ^condition))
+            {ref, state}
 
-    {ref, state}
+          :error ->
+            :error
+        end
+
+      :error ->
+        :error
+    end
   end
 
   defp build(value, state) do
@@ -363,46 +473,206 @@ defmodule Dx.Ecto.Scope do
     |> with_state(state)
   end
 
-  defp build_condition(fun, base, state) when is_function(fun, 1) do
-    base |> fun.() |> build_condition(base, state)
+  defp build_query_filter(state, :skip), do: state
+  defp build_query_filter(state, condition), do: map_query(state, &where(&1, ^condition))
+
+  defp build_fallback_filter(state, :skip), do: state
+
+  defp build_fallback_filter(state, fallback) do
+    Map.update!(state, :post_load, &{:filter, fallback, {:ref, current_ref(state)}, &1})
   end
 
-  defp build_condition({:error, defd_fun}, _base, state) do
-    %{queries: [%{ref: base_ref} | _queries]} = state
-
-    state = Map.update!(state, :post_load, &{:filter, defd_fun, {:ref, base_ref}, &1})
-
-    {true, state}
+  # build_condition/2 can return either
+  #   {query, fallback, state}
+  #     query :: Ecto dynamic | :skip
+  #     fallback :: fun | :skip | {:all_of, funs} | {:any_of, funs}
+  #   :error (no fallback possible)
+  #
+  defp build_condition({:error, defd_fun}, state) do
+    {:skip, defd_fun, state}
   end
 
-  defp build_condition({:all_of, conditions}, base, state) do
-    {conditions, state} = Enum.map_reduce(conditions, state, &build_condition(&1, base, &2))
-    conditions = Enum.reduce(conditions, &dynamic(^&2 and ^&1))
+  defp build_condition({:all_of, conditions}, state) do
+    Enum.reduce_while(conditions, {[], [], state}, fn condition, {conditions, fallbacks, state} ->
+      case build_condition(condition, state) do
+        {condition, :skip, state} ->
+          {:cont, {maybe_prepend(condition, conditions), fallbacks, state}}
 
-    {conditions, state}
+        {_condition, _fallback, state} when in_subquery?(state) ->
+          {:halt, :error}
+
+        {condition, fallback, state} ->
+          {:cont,
+           {maybe_prepend(condition, conditions), maybe_prepend(fallback, fallbacks), state}}
+
+        :error when in_subquery?(state) ->
+          {:halt, :error}
+
+        :error ->
+          raise ArgumentError,
+                "Condition must provide fallback function: \n\n" <>
+                  inspect(condition, pretty: true)
+      end
+    end)
+    |> case do
+      {conditions, fallbacks, state} ->
+        conditions =
+          case conditions do
+            [] -> :skip
+            conditions -> conditions |> :lists.reverse() |> Enum.reduce(&dynamic(^&2 and ^&1))
+          end
+
+        fallbacks =
+          case fallbacks do
+            [] -> :skip
+            [fallback] -> fallback
+            fallbacks -> {:all_of, fallbacks}
+          end
+
+        {conditions, fallbacks, state}
+
+      :error ->
+        :error
+    end
   end
 
-  defp build_condition({:not, condition}, base, state) do
-    {condition, state} = build_condition(condition, base, state)
+  defp build_condition({:any_of, conditions, fallback}, state) do
+    Enum.reduce_while(conditions, {[], [], state}, fn condition, {conditions, fallbacks, state} ->
+      case build_condition(condition, state) do
+        {:skip, _, _} ->
+          {:halt, :skip}
 
-    dynamic(not (^condition))
-    |> with_state(state)
+        {condition, fallback, state} ->
+          {:cont,
+           {maybe_prepend(condition, conditions), maybe_prepend(fallback, fallbacks), state}}
+
+        :error ->
+          {:halt, :error}
+      end
+    end)
+    |> case do
+      {conditions, fallbacks, state} ->
+        conditions =
+          case conditions do
+            [] -> :skip
+            conditions -> conditions |> :lists.reverse() |> Enum.reduce(&dynamic(^&2 or ^&1))
+          end
+
+        fallbacks =
+          case fallbacks do
+            [] -> :skip
+            _other -> fallback
+          end
+
+        {conditions, fallbacks, state}
+
+      :skip ->
+        {:skip, fallback, state}
+
+      :error ->
+        :error
+    end
   end
 
-  defp build_condition({:eq, left, nil}, _base, state) do
-    {left, state} = build_value(left, state)
+  defp build_condition({:any?, base, fallback}, state) do
+    case build(base, state) do
+      {_ref, %{queries: [%{query: subquery} | queries]} = state} ->
+        state = %{state | queries: queries}
+        subquery = select(subquery, true)
 
-    dynamic(is_nil(^left))
-    |> with_state(state)
+        dynamic(exists(subquery(subquery)))
+        |> with_fallback_and_state(state)
+
+      :error when in_subquery?(state) ->
+        :error
+
+      :error ->
+        {:skip, fallback, state}
+    end
   end
 
-  defp build_condition({:eq, left, right}, _base, state) do
-    {left, state} = build_value(left, state)
-    {right, state} = build_value(right, state)
+  defp build_condition({:not, condition}, state) do
+    case build_condition(condition, state) do
+      {condition, :skip, state} ->
+        dynamic(not (^condition))
+        |> with_fallback_and_state(state)
 
-    dynamic(^left == ^right)
-    |> with_state(state)
+      {:skip, fallback, state} ->
+        {:skip, {:not, fallback}, state}
+
+      {condition, fallback, state} ->
+        dynamic(not (^condition))
+        |> with_fallback_and_state({:not, fallback}, state)
+
+      :error ->
+        :error
+    end
   end
+
+  defp build_condition({:eq, left, nil}, state) do
+    case build_value(left, state) do
+      {left, state} ->
+        dynamic(is_nil(^left))
+        |> with_fallback_and_state(state)
+
+      :error ->
+        :error
+    end
+  end
+
+  defp build_condition({:eq, left, right}, state) do
+    case build_value(left, state) do
+      {left, state} ->
+        case build_value(right, state) do
+          {right, state} ->
+            dynamic(^left == ^right)
+            |> with_fallback_and_state(state)
+
+          :error ->
+            :error
+        end
+
+      :error ->
+        :error
+    end
+  end
+
+  defp build_condition({:lt, left, right}, state) do
+    case build_value(left, state) do
+      {left, state} ->
+        case build_value(right, state) do
+          {right, state} ->
+            dynamic(^left < ^right)
+            |> with_fallback_and_state(state)
+
+          :error ->
+            :error
+        end
+
+      :error ->
+        :error
+    end
+  end
+
+  defp build_condition({:gt, left, right}, state) do
+    case build_value(left, state) do
+      {left, state} ->
+        case build_value(right, state) do
+          {right, state} ->
+            dynamic(^left > ^right)
+            |> with_fallback_and_state(state)
+
+          :error ->
+            :error
+        end
+
+      :error ->
+        :error
+    end
+  end
+
+  defp maybe_prepend(:skip, acc), do: acc
+  defp maybe_prepend(elem, acc), do: [elem | acc]
 
   defp build_value(term, state) do
     case build(term, state) do
@@ -414,9 +684,10 @@ defmodule Dx.Ecto.Scope do
     end
   end
 
-  def run_post_load(results, {:filter, fun, {:ref, :a0}, {:loaded}}, eval) do
+  def run_post_load(results, {:filter, {:not, fun}, {:ref, :a0}, rest}, eval) do
     results
-    |> Dx.Defd.Result.then(&Dx.Defd.Result.filter(&1, fun, eval))
+    |> run_post_load(rest, eval)
+    |> Dx.Defd.Result.then(&Dx.Defd.Result.reject(&1, fun, eval))
   end
 
   def run_post_load(results, {:filter, fun, {:ref, :a0}, rest}, eval) do
@@ -442,8 +713,17 @@ defmodule Dx.Ecto.Scope do
 
   ## Helpers
 
-  @compile {:inline, with_state: 2, map_query: 2, map_query: 3}
+  @compile {:inline,
+            with_state: 2,
+            with_fallback_and_state: 2,
+            with_fallback_and_state: 3,
+            current_ref: 1,
+            map_query: 2,
+            map_query: 3}
   defp with_state(term, state), do: {term, state}
+  defp with_fallback_and_state(term, fallback \\ :skip, state), do: {term, fallback, state}
+
+  defp current_ref(%{queries: [%{ref: ref} | _queries]}), do: ref
 
   defp map_query(%{queries: [query | queries]} = state, fun) do
     %{state | queries: [Map.update!(query, :query, fun) | queries]}